pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-distill-mcp"
version = "0.1.2"
description = "MCP server for academic paper search, curation, and multi-platform push"
license = "AGPL-3.0-only"
requires-python = ">=3.10"
readme = "README.md"
authors = [
    { name = "Jie Cao", email = "vertex.cj@gmail.com" },
]
keywords = ["mcp", "paper", "academic", "search", "arxiv", "pubmed", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.0.0",
    "arxiv>=2.0.0",
    "biopython>=1.80",
    "pyzotero>=1.5.0",
    "python-telegram-bot>=22.0",
    "openai>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.urls]
Homepage = "https://github.com/Eclipse-Cj/paper-distill-mcp"
Repository = "https://github.com/Eclipse-Cj/paper-distill-mcp"
Issues = "https://github.com/Eclipse-Cj/paper-distill-mcp/issues"

[project.scripts]
paper-distill-mcp = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "search", "curate", "generate", "integrations", "bot", "paper_digest"]
